% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitGompertz.R
\name{FitGompertz}
\alias{FitGompertz}
\title{Fit a Gompertz function to 'Slakes' data}
\usage{
FitGompertz(SlakesData = NULL, path = NULL, id = NULL)
}
\arguments{
\item{SlakesData}{data obtained by the function \code{\link{readSlakes}}}
}
\description{
This fits a Gompertz-like function on a specific range of values (same than the android app)
using Area versus Time data imported by the function \code{\link{readSlakes}}
}
\author{
Mario Fajardo
}

