% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSlakes.R
\name{plotSlakes}
\alias{plotSlakes}
\title{Plot 'slakes' data}
\usage{
plotSlakes(dataSlakes = NULL, logScale = TRUE, facet = "Treatment",
  Id = "ALL")
}
\arguments{
\item{dataSlakes}{data obtained by using the function \code{\link{readSlakes}}}
}
\value{
A \code{\link{ggplot}} object
}
\description{
This function creates a \code{\link{ggplot}} object and plots data exported from the android app "slakes"
}
\author{
Mario Fajardo
}

