% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoefs.R
\name{readCoefs}
\alias{readCoefs}
\title{Read 'slakes' coefficients data}
\usage{
readCoefs(path = NULL)
}
\arguments{
\item{path}{path to the folder "slakes"}
}
\description{
This function imports soil slaking fitted coefficients from the android app "slakes"
}
\author{
Mario Fajardo
}

