% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSlakes.R
\name{readSlakes}
\alias{readSlakes}
\title{Read 'slakes' data}
\usage{
readSlakes(path = NULL)
}
\arguments{
\item{path}{path to the folder "slakes"}
}
\description{
This function imports data exported from the android app "slakes"
}
\author{
Mario Fajardo
}

